Refactor Gizmos system and expose DrawGizmos in Behaviour
Replaces GizmosModule with a new Gizmos class, updates all references and usage accordingly, and removes the old GizmosModule files. Adds a DrawGizmos virtual method to Behaviour, allowing behaviours to render custom gizmos. Updates RotateAroundAxis to implement DrawGizmos, and ensures DefaultRenderScenario calls RenderBehaviourGizmos. Also makes minor improvements to Vector3 API, input handling, and replaces math::lerp with std::lerp in Color.

+++ Rei/src/Modules/Editor/FlyCameraSystem.cpp
@@ -31,7 +31,7 @@ void rei::editor::FlyCameraSystem::OnUpdate()
 
     if (!Input::IsMouseButtonDown(GLFW_MOUSE_BUTTON_RIGHT)) return;
 
-    float currentFrame = glfwGetTime();
+    double currentFrame = glfwGetTime();
     deltaTime = currentFrame - lastFrame;
     lastFrame = currentFrame;
